splitup persistent_mapset_merger
